requirements.txt
gunicorn
flask
dash
plotly
pandas
python-dateutil
flask-heroku
flask-sqlalchemy
flask-migrate
psycopg2-binary